Update delete-bad-users.js
Not updates users with bad lids if user name lines up with something that exists.

+++ tools/delete-bad-users.js
@@ -1,20 +1,21 @@
 #!/usr/bin/env node
 
-const ConfigFactory = require('../factories/config-factory');
-const DependencyInjectionContainer = require('../dependency-injection-container');
-const dotenv = require('dotenv');
-const fs = require('fs');
+// Allow switching between dev and prod
+process.env.npm_config_live = process.argv[2] == 'prod' || false;
+const config = require('../config.js');
+if (process.env.npm_config_live) {
+    config.set('googleCloudIdentityKeyFile', './.gcp-key.json');
+}
 
-// Load .env into process.env, create config, create container
-dotenv.config();
-const configModel = new ConfigFactory('prod', process.env, {}, fs.existsSync).build();
-const container = new DependencyInjectionContainer(configModel);
-
+// ...and go!
+const container = require('../dependency-injection-container')(config);
+const httpClient = container.resolve('httpClient');
 const collection = container.resolve('firestoreConnection').getCollection();
 processData(collection);
 
 async function processData(collection) {
-    const querySnapshot = await collection.get();
+    // hard code some limits here so I can do the process in steps
+    const querySnapshot = await collection.orderBy('userName').startAt('a').endAt('b').get();
     var documentList = querySnapshot.docs;
     const fileName = 'Bad User Deletes ' + new Date().toString() + '.txt';
 
@@ -22,12 +23,13 @@ async function processData(collection) {
         console.log(`${documentList.length} users to check...`);
         for (const key in documentList) {
             const data = documentList[key].data();
-            try {
-                await container.resolve('letterboxdDiaryRss').get(data.userName, 1);
-                console.log(`+++ Success on ${data.userName}`);
+
+            const validLetterBoxdId = await getUserLetterboxdStatus(data.letterboxdId);
+            if (validLetterBoxdId) {
                 documentList[key] = false;
-            } catch (error) {
-                console.log(`--- Failure on ${data.userName}`);
+                console.log(`👉 ${data.userName}/${data.letterboxdId} validated via user lid`);
+            } else {
+                console.log(`❌ ${data.userName}/${data.letterboxdId} not validated via user lid`);
             }
         }
         documentList = documentList.filter((x) => !!x);
@@ -35,10 +37,57 @@ async function processData(collection) {
 
     for (const key in documentList) {
         const document = documentList[key];
+        const id = documentList[key].id;
         const data = document.data();
 
-        const content = data.userName + '||' + JSON.stringify(data.channelList) + '\n';
-        fs.writeFileSync(fileName, content, { flag: 'a+' });
-        await document.ref.delete();
+        let userLid = await getUserNameStatus(data.userName);
+        if (userLid == '') {
+            console.log(`❌ ${data.userName} not able to validate user name`);
+            userLid = await getUserNameStatus(id);
+        }
+
+        if (userLid) {
+            const realUserName = await getUserLetterboxdStatus(userLid);
+            console.log(`✅ Updating record for /${id}/ with ${realUserName}/${userLid}`);
+
+            data.letterboxdId = userLid;
+            data.userName = realUserName;
+
+            const documentSnapshot = querySnapshot.docs[key];
+            await documentSnapshot.ref.update(data);
+        } else {
+            console.log(`❌ /${id}/ not able to validate with document key`);
+            await document.ref.delete();
+        }
     }
+}
+
+async function getUserLetterboxdStatus(letterboxdId) {
+    return httpClient
+        .head('https://boxd.it/' + letterboxdId, 10000)
+        .then((response) => {
+            if (response?.headers['x-letterboxd-type'] != 'Member') {
+                throw 'error';
+            }
+
+            const url = response?.request?.path;
+            return url.replaceAll('/', '');
+        })
+        .catch(() => {
+            return '';
+        });
+}
+
+async function getUserNameStatus(userName) {
+    return httpClient
+        .head('https://letterboxd.com/' + userName, 10000)
+        .then((response) => {
+            if (response?.headers['x-letterboxd-type'] != 'Member') {
+                throw 'error';
+            }
+            return response?.headers['x-letterboxd-identifier'] || '';
+        })
+        .catch(() => {
+            return '';
+        });
 }
